Rename variables and methods in system tests to use file reference terminology
